Refactor: Repository Pattern으로 변경
